fix e2e on esm. switch from ts-node to tsx. minor import fixes.
