+++ apps/e2e/src/engine-client/fullSanity.ts
@@ -11,7 +11,7 @@ import {
   VERTEX_ABIS,
   VLP_PRODUCT_ID,
 } from '@vertex-protocol/contracts';
-import { MOCK_ERC20_ABI } from '@vertex-protocol/contracts/dist/common/abis/MockERC20';
+import { MOCK_ERC20_ABI } from '@vertex-protocol/contracts';
 import {
   EngineClient,
   EngineIsolatedOrderParams,
