+++ package.json
@@ -37,7 +37,7 @@
     "lint-staged": "^15.2.7",
     "prettier": "^3.4.1",
     "ts-jest": "^29.2.5",
-    "ts-node": "^10.9.1",
+    "tsx": "4.19.4",
     "typedoc": "^0.26.5",
     "typedoc-plugin-extras": "^3.1.0",
     "typedoc-plugin-missing-exports": "^3.0.0",
